Avance - agregado modulo de inventarios - agregado modulo de resumen de cierre de inventario actual - agregado modulo de cierres de inventarios - agregado registro de inventarios - agregado cierre de inventarios - agregado exportacion de cierres de inventarios e inventarios - agregado modulo de edicion de catalogos de formulario de inventarios - agregado dtos de inventarios, cierres y catalogos

+++ GComFuelManager/Shared/Modelos/Catalogo_Fijo.cs
@@ -12,7 +12,7 @@ namespace GComFuelManager.Shared.Modelos
         public int Id { get; set; }
         [StringLength(250)]
         public string Valor { get; set; } = string.Empty;
-        public short Catalogo { get; set; }
+        public int Catalogo { get; set; }
         public bool Activo { get; set; } = true;
         public string? Abreviacion { get; set; } = string.Empty;
